requirements.txt
aniso8601==7.0.0
asgiref==3.3.1
autopep8==1.5.6
Django==3.1.7
django-cors-headers==3.7.0
django-graphql-jwt==0.3.0
graphene==2.1.8
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
promise==2.3
pycodestyle==2.7.0
PyJWT==1.7.0
pytz==2021.1
Rx==1.6.1
singledispatch==3.6.1
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
